+++ samples/bottle-sample.py
@@ -2,8 +2,8 @@ import sys, os
 from bottle import run, request, route
 
 # Not ideal, but for the sake of an example
-sys.path.append(os.path.join(os.path.dirname(os.path.dirname(__file__)), 'src'))
-from steamsignin import SteamSignIn
+sys.path.append(os.path.join(os.path.dirname(os.path.dirname(__file__))))
+from pySteamSignIn.steamsignin import SteamSignIn
 
 @route('/')
 def main():
